fix(ui): register the 'bot' icon and fail the build on registry drift (#5637)
`channel_discord`'s AI-proposal message type declares `icon: 'bot'`, but the
committed `lucideRegistry.generated.tsx` had no `'bot'` key, so
`resolveRegisteredLucideIcon` missed and the message type rendered with no icon
at all.

The registry is build output: `generateLucideRegistry()` scans `apps/**` and
`packages/**` for `icon: '…'` / `icon="…"` literals, resolves each kebab name to
a `lucide-react` export, and writes the file. `Bot` is a real lucide export and
the generator does pick it up — the `channel-discord` module simply landed
without the regenerated registry. Regenerating produces exactly the two lines
that were missing, taking the registry from 160 to 161 icons.

Nothing caught it, and nothing would have caught the next one. The existing
suite exercises the *emitter* against hand-written fixtures and never compares
the *committed* file against what a scan of the repo would produce, so an icon
reference can merge while the registry that resolves it stays stale — a silent
failure, since a missing key renders nothing rather than throwing.

So the discovery pass moves out of the inline body of `build.mjs` into
`scripts/lucideIconDiscovery.cjs`, side-effect free and CommonJS for the same
reason `lucideRegistrySource.cjs` already is: both the build and the Jest suite
need to load it. The ignore list, both regex patterns and the `Icon` exclusion
move across verbatim — a real `yarn build:packages` reproduces the file
byte-for-byte, which is what pins the extraction as behaviour-preserving.

Two tests then close the loop: one asserts every discovered icon string is
registered and names the ones that are not, the other asserts the committed file
is byte-identical to a fresh generator run, catching stale extras and formatting
drift too. Both test names carry the remediation command, so a CI failure
explains its own fix. Stashing only the regenerated registry turns both red,
reporting exactly `"bot"`. The repo-wide scan adds no meaningful runtime — the
icons suites still finish in ~3.4s.
