+++ packages/ui/src/backend/icons/lucideRegistry.generated.tsx
@@ -22,6 +22,7 @@ import {
   Bolt,
   Book,
   Bookmark,
+  Bot,
   Box,
   Boxes,
   Briefcase,
@@ -185,6 +186,7 @@ export const LUCIDE_ICON_REGISTRY: Record<string, LucideIcon> = {
   'bolt': Bolt,
   'book': Book,
   'bookmark': Bookmark,
+  'bot': Bot,
   'box': Box,
   'boxes': Boxes,
   'briefcase': Briefcase,
